requirements.txt
flask
gunicorn
resend
psycopg2-binary
